feat(signature-refactor): review migration file and create model (#5151)
