+++ openaev-api/src/main/java/io/openaev/utils/InjectExpectationSignatureUtils.java
@@ -1,0 +1,61 @@
+package io.openaev.utils;
+
+import io.openaev.database.model.InjectExpectationSignature;
+import io.openaev.validator.Ipv4OrIpv6Validator;
+
+public class InjectExpectationSignatureUtils {
+  public static final String EXPECTATION_SIGNATURE_TYPE_PARENT_PROCESS_NAME = "parent_process_name";
+  public static final String EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV4_ADDRESS = "source_ipv4_address";
+  public static final String EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV6_ADDRESS = "source_ipv6_address";
+  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_IPV4_ADDRESS = "target_ipv4_address";
+  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_IPV6_ADDRESS = "target_ipv6_address";
+  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_HOSTNAME_ADDRESS =
+      "target_hostname_address";
+  public static final String EXPECTATION_SIGNATURE_TYPE_START_DATE = "start_date";
+  public static final String EXPECTATION_SIGNATURE_TYPE_END_DATE = "end_date";
+  // AI adversarial validation: correlate AI defense (LLM firewall / guardrail) events back to a
+  // specific AI inject execution. The injector emits a unique per-inject marker and the target
+  // endpoint it called; AI defense collectors match these against their logs.
+  public static final String EXPECTATION_SIGNATURE_TYPE_AI_REQUEST_MARKER = "ai_request_marker";
+  public static final String EXPECTATION_SIGNATURE_TYPE_AI_TARGET_ENDPOINT = "ai_target_endpoint";
+
+  public static InjectExpectationSignature createIpSignature(String ip, boolean isTarget) {
+    if (ip == null || ip.isEmpty()) {
+      return null;
+    }
+    if (Ipv4OrIpv6Validator.isIpv4(ip)) {
+      return new InjectExpectationSignature(
+          isTarget
+              ? EXPECTATION_SIGNATURE_TYPE_TARGET_IPV4_ADDRESS
+              : EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV4_ADDRESS,
+          ip);
+    } else if (Ipv4OrIpv6Validator.isIpv6(ip)) {
+      return new InjectExpectationSignature(
+          isTarget
+              ? EXPECTATION_SIGNATURE_TYPE_TARGET_IPV6_ADDRESS
+              : EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV6_ADDRESS,
+          ip);
+    } else {
+      return null;
+    }
+  }
+
+  public static InjectExpectationSignature createHostnameSignature(String signatureValue) {
+    return new InjectExpectationSignature(
+        EXPECTATION_SIGNATURE_TYPE_TARGET_HOSTNAME_ADDRESS, signatureValue);
+  }
+
+  public static InjectExpectationSignature createAiRequestMarkerSignature(String marker) {
+    if (marker == null || marker.isEmpty()) {
+      return null;
+    }
+    return new InjectExpectationSignature(EXPECTATION_SIGNATURE_TYPE_AI_REQUEST_MARKER, marker);
+  }
+
+  public static InjectExpectationSignature createAiTargetEndpointSignature(String endpoint) {
+    if (endpoint == null || endpoint.isEmpty()) {
+      return null;
+    }
+    return new InjectExpectationSignature(EXPECTATION_SIGNATURE_TYPE_AI_TARGET_ENDPOINT, endpoint);
+  }
+}
