+++ openaev-model/src/main/java/io/openaev/database/model/InjectExpectationSignature.java
@@ -1,72 +1,57 @@
 package io.openaev.database.model;
 
-import io.openaev.validator.Ipv4OrIpv6Validator;
-import lombok.Builder;
-import lombok.Data;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
 
-@Data
+import java.io.Serializable;
+import java.time.Instant;
+
+import static java.time.Instant.now;
+
+@Getter
+@Setter
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(of = {"type", "value"})
+@Entity
+@IdClass(InjectExpectationSignature.InjectExpectationSignatureId.class)
+@Table(name = "injects_expectations_signatures")
 public class InjectExpectationSignature {
-  public static final String EXPECTATION_SIGNATURE_TYPE_PARENT_PROCESS_NAME = "parent_process_name";
-  public static final String EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV4_ADDRESS = "source_ipv4_address";
-  public static final String EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV6_ADDRESS = "source_ipv6_address";
-  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_IPV4_ADDRESS = "target_ipv4_address";
-  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_IPV6_ADDRESS = "target_ipv6_address";
-  public static final String EXPECTATION_SIGNATURE_TYPE_TARGET_HOSTNAME_ADDRESS =
-      "target_hostname_address";
-  public static final String EXPECTATION_SIGNATURE_TYPE_START_DATE = "start_date";
-  public static final String EXPECTATION_SIGNATURE_TYPE_END_DATE = "end_date";
-  // AI adversarial validation: correlate AI defense (LLM firewall / guardrail) events back to a
-  // specific AI inject execution. The injector emits a unique per-inject marker and the target
-  // endpoint it called; AI defense collectors match these against their logs.
-  public static final String EXPECTATION_SIGNATURE_TYPE_AI_REQUEST_MARKER = "ai_request_marker";
-  public static final String EXPECTATION_SIGNATURE_TYPE_AI_TARGET_ENDPOINT = "ai_target_endpoint";
 
+  @Id
+  @JsonIgnore
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "inject_expectation_signature_inject_expectation_id")
+  private InjectExpectation injectExpectation;
+
+  @Id
+  @NotBlank
+  @Column(name = "inject_expectation_signature_type")
+  @JsonProperty("type")
   private String type;
+
+  @NotBlank
+  @Column(name = "inject_expectation_signature_value")
+  @JsonProperty("value")
   private String value;
 
-  public InjectExpectationSignature(String signatureType, String signatureValue) {
-    this.type = signatureType;
-    this.value = signatureValue;
-  }
+  @NotNull
+  @JsonIgnore
+  @Column(name = "inject_expectation_created_at", updatable = false)
+  @JsonProperty("inject_expectation_signature_created_at")
+  private Instant createdAt = now();
 
-  public static InjectExpectationSignature createIpSignature(String ip, boolean isTarget) {
-    if (ip == null || ip.isEmpty()) {
-      return null;
-    }
-    if (Ipv4OrIpv6Validator.isIpv4(ip)) {
-      return new InjectExpectationSignature(
-          isTarget
-              ? EXPECTATION_SIGNATURE_TYPE_TARGET_IPV4_ADDRESS
-              : EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV4_ADDRESS,
-          ip);
-    } else if (Ipv4OrIpv6Validator.isIpv6(ip)) {
-      return new InjectExpectationSignature(
-          isTarget
-              ? EXPECTATION_SIGNATURE_TYPE_TARGET_IPV6_ADDRESS
-              : EXPECTATION_SIGNATURE_TYPE_SOURCE_IPV6_ADDRESS,
-          ip);
-    } else {
-      return null;
-    }
-  }
-
-  public static InjectExpectationSignature createHostnameSignature(String signatureValue) {
-    return new InjectExpectationSignature(
-        EXPECTATION_SIGNATURE_TYPE_TARGET_HOSTNAME_ADDRESS, signatureValue);
-  }
-
-  public static InjectExpectationSignature createAiRequestMarkerSignature(String marker) {
-    if (marker == null || marker.isEmpty()) {
-      return null;
-    }
-    return new InjectExpectationSignature(EXPECTATION_SIGNATURE_TYPE_AI_REQUEST_MARKER, marker);
-  }
-
-  public static InjectExpectationSignature createAiTargetEndpointSignature(String endpoint) {
-    if (endpoint == null || endpoint.isEmpty()) {
-      return null;
-    }
-    return new InjectExpectationSignature(EXPECTATION_SIGNATURE_TYPE_AI_TARGET_ENDPOINT, endpoint);
+  @Getter
+  @Setter
+  @NoArgsConstructor
+  @AllArgsConstructor
+  public static class InjectExpectationSignatureId implements Serializable {
+    private String injectExpectation;
+    private String type;
   }
 }
